Restructure tests for multiple hyperscript implementations
